Create new package, only with configuration classes, added class for replases DispatcherServlet-context.xml (now SpringConfiguration.java) and web.xml (now InitApplication.java)
