+++ pom.xml
@@ -15,6 +15,23 @@
                     <target>1.7</target>
                 </configuration>
             </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-war-plugin</artifactId>
+                <version>2.1.1</version>
+                <configuration>
+                    <failOnMissingWebXml>false</failOnMissingWebXml>
+                </configuration>
+            </plugin>
+            <plugin>
+                 <groupId>org.apache.tomcat.maven</groupId>
+                 <artifactId>tomcat7-maven-plugin</artifactId>
+                 <version>2.0</version>
+                 <configuration>
+                      <port>8080</port>
+                      <path>/</path>
+                 </configuration>
+            </plugin>
         </plugins>
     </build>
     <packaging>war</packaging>
