+++ src/main/java/com/shop/configuration/InitApplication.java
@@ -1,0 +1,26 @@
+package com.shop.configuration;
+
+import org.springframework.web.WebApplicationInitializer;
+import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
+import org.springframework.web.servlet.DispatcherServlet;
+
+import javax.servlet.ServletContext;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRegistration;
+
+/**
+ * Created by RSzczygielski on 12.02.16.
+ */
+public class InitApplication implements WebApplicationInitializer {
+    @Override
+    public void onStartup(ServletContext servletContext) throws ServletException {
+        AnnotationConfigWebApplicationContext context = new AnnotationConfigWebApplicationContext();
+
+        context.register(SpringConfiguration.class);
+        DispatcherServlet dispatcherServlet = new DispatcherServlet(context);
+        ServletRegistration.Dynamic servletRegistration = servletContext.addServlet("DispatcherServlet", dispatcherServlet);
+
+        servletRegistration.addMapping("/");
+        servletRegistration.setLoadOnStartup(1);
+    }
+}
