+++ src/main/java/com/shop/configuration/SpringConfiguration.java
@@ -1,0 +1,47 @@
+package com.shop.configuration;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.ResourceBundleMessageSource;
+import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
+import org.springframework.web.servlet.ViewResolver;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+import org.springframework.web.servlet.view.InternalResourceViewResolver;
+import org.springframework.web.servlet.view.JstlView;
+
+/**
+ * Created by RSzczygielski on 12.02.16.
+ */
+@Configuration
+@EnableWebMvc
+@ComponentScan("com.shop")
+public class SpringConfiguration extends WebMvcConfigurerAdapter{
+    @Bean
+    public ViewResolver viewResolver() {
+        InternalResourceViewResolver resourceViewResolver = new InternalResourceViewResolver();
+        resourceViewResolver.setViewClass(JstlView.class);
+        resourceViewResolver.setPrefix("/WEB-INF/jsp/");
+        resourceViewResolver.setSuffix(".jsp");
+
+        return resourceViewResolver;
+    }
+
+    @Bean
+    public ResourceBundleMessageSource messageSource() {
+        ResourceBundleMessageSource bundleMessageSource = new ResourceBundleMessageSource();
+        String[] filesWithMessages = {"messages/static", "messages/validator"};
+        bundleMessageSource.setBasenames(filesWithMessages);
+
+        return bundleMessageSource;
+    }
+
+    @Bean
+    public LocalValidatorFactoryBean getValidator() {
+        LocalValidatorFactoryBean lvfb = new LocalValidatorFactoryBean();
+        lvfb.setValidationMessageSource(messageSource());
+
+        return lvfb;
+    }
+}
